Remove package imports from .d.ts file
